version 0.2
I reworked the home page to have a table display the statistics in a way that is more visually appealing. I edited the edit_player page to give an option on if you want to edit an existing player or add a new one. I changed the main app.py file to work with those new layouts. I added a navbar.html file and added that to each of the pages.